refactor(active-info): 활동 정보 API를 다른 파일로 분리한다
