+++ src/mocks/handlers/index.ts
@@ -1,3 +1,4 @@
+import * as activeInfoHandlers from '@/mocks/handlers/active-info';
 import * as authHandlers from '@/mocks/handlers/auth';
 import * as panelHandlers from '@/mocks/handlers/panel';
 import * as questionHandlers from '@/mocks/handlers/question';
@@ -6,4 +7,5 @@ export const handlers = [
   ...Object.values(authHandlers),
   ...Object.values(panelHandlers),
   ...Object.values(questionHandlers),
+  ...Object.values(activeInfoHandlers),
 ];
