+++ src/mocks/handlers/panel.ts
@@ -9,19 +9,13 @@ import type {
   GetMyPanelsResponse,
   GetPanelPathParams,
   GetPanelResponse,
-  GetMyActiveInfoResponse,
-  GetMyActiveInfoPathParams,
 } from '@/lib/api/panel';
 
 import { rest } from 'msw';
 
 import { apiUrl } from '@/lib/api/apiUrl';
 import { API_BASE_URL } from '@/lib/apiClient';
-import {
-  createMockPanel,
-  myActiveInfoMock,
-  myPanelList,
-} from '@/mocks/data/panel';
+import { createMockPanel, myPanelList } from '@/mocks/data/panel';
 
 import { myAccount } from '../data/auth';
 
@@ -126,27 +120,6 @@ export const getPanel = rest.get<never, GetPanelPathParams, GetPanelResponse>(
       ctx.json({
         statusCode: 200,
         result: panel,
-      }),
-    );
-  },
-);
-
-export const getMyActiveInfo = rest.get<
-  never,
-  GetMyActiveInfoPathParams,
-  GetMyActiveInfoResponse
->(
-  `${API_BASE_URL}${apiUrl.panel.getMyActiveInfo(':panelId')}`,
-  async (req, res, ctx) => {
-    let activeInfoToken = req.cookies.active_info_id;
-
-    if (!activeInfoToken?.length) activeInfoToken = 'dev-active-info-token';
-    return res(
-      ctx.status(200),
-      ctx.cookie('active_info_id', activeInfoToken, { httpOnly: true }),
-      ctx.json({
-        statusCode: 200,
-        result: myActiveInfoMock,
       }),
     );
   },
